+++ packages/hoppscotch-common/src/services/persistence/__tests__/__mocks__/index.ts
@@ -116,9 +116,10 @@ export const MQTT_REQUEST_MOCK = {
   clientID: "hoppscotch",
 }
 
-export const GLOBAL_ENV_MOCK: GlobalEnvironment["variables"] = [
-  { key: "test-key", value: "test-value", secret: false },
-]
+export const GLOBAL_ENV_MOCK: GlobalEnvironment = {
+  v: 1,
+  variables: [{ key: "test-key", value: "test-value", secret: false }],
+}
 
 export const VUEX_DATA_MOCK: VUEX_DATA = {
   postwoman: {
