+++ packages/hoppscotch-common/src/services/persistence/index.ts
@@ -628,7 +628,7 @@ export class PersistenceService extends Service {
   private setupGlobalEnvsPersistence() {
     const globalEnvKey = "globalEnv"
     let globalEnvData: GlobalEnvironment = JSON.parse(
-      window.localStorage.getItem(globalEnvKey) || "{}"
+      window.localStorage.getItem(globalEnvKey) || "[]"
     )
 
     // Validate data read from localStorage
