+++ packages/hoppscotch-data/src/index.ts
@@ -3,4 +3,4 @@ export * from "./graphql"
 export * from "./collection"
 export * from "./rawKeyValue"
 export * from "./environment"
-export * from "./global-environments"
+export * from "./global-environment"
